refactor: 구조 최적화
