+++ public/sw.js
@@ -1,1 +1,1 @@
-if(!self.define){let e,s={};const i=(i,a)=>(i=new URL(i+".js",a).href,s[i]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=i,e.onload=s,document.head.appendChild(e)}else e=i,importScripts(i),s()})).then((()=>{let e=s[i];if(!e)throw new Error(`Module ${i} didn’t register its module`);return e})));self.define=(a,n)=>{const t=e||("document"in self?document.currentScript.src:"")||location.href;if(s[t])return;let c={};const o=e=>i(e,t),r={module:{uri:t},exports:c,require:o};s[t]=Promise.all(a.map((e=>r[e]||o(e)))).then((e=>(n(...e),c)))}}define(["./workbox-e9849328"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"0e4cd552c0d7c364bedd388c2039eafd"},{url:"/_next/static/b8k8igtsSyX_QFxIY3ZiQ/_buildManifest.js",revision:"a9fb6c6c2b196342b346f13620ff8e60"},{url:"/_next/static/b8k8igtsSyX_QFxIY3ZiQ/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/151-fb9977fa774b1fb0.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/281-98e2740ba6a70311.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/341-707724d4b222fb41.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/465-997053893adc2514.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/515-97cf8580fd196a8d.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/523-39a8b1d5137fe307.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/594-82244899bd5374a8.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/687-5a5e31213e1abb8d.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/705-b4a7d38f16abd233.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/728-10893fcf8fe80b2b.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/997-82fa72b95b599b2a.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/app/(auth)/policy/page-e859b46aab4f6ba6.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/app/(auth)/sign-in/page-ec14319cb91c02b0.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/app/(auth)/sign-up-complete/page-b19a897f2755284c.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/app/(auth)/sign-up/page-406054075f2bf6db.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/app/(main)/new-content/layout-8f354e0345b35c55.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/app/(main)/new-content/page-439c170288819329.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/app/(modal)/warning/page-e7c2250c2b09c4a8.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/app/@modal/(.)warning/page-234df44c1392f264.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/app/@modal/default-faae72e4101ecadb.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/app/_not-found/page-c47ff0b24eac5816.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/app/error-cae605b8839498b7.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/app/layout-3df81cf7c965abe4.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/app/loading-bdcb4e5f0b2cb86d.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/app/not-found-8771b6be3a4e8aa2.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/app/onboarding/page-913b7e53305a7609.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/app/page-cc3760f5a2cbd47e.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/ca777b1b-1c1ec62045d69b42.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/framework-6e06c675866dc992.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/main-3f632235c87e07b4.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/main-app-063f3418fe6187ac.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/pages/_app-399de83ec01c3392.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/pages/_error-82c8287d732697b1.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-70b9a19db61b851d.js",revision:"b8k8igtsSyX_QFxIY3ZiQ"},{url:"/_next/static/css/291a1637afd87740.css",revision:"291a1637afd87740"},{url:"/_next/static/media/5d85a1edd96c32f7-s.p.woff2",revision:"fae6df7521d3c7a4978234c4311ccb2b"},{url:"/_next/static/media/9edda583dedd31eb-s.p.woff2",revision:"100caa3008af382762a57d6e0e9d2673"},{url:"/_next/static/media/e27282d13bd93057-s.p.woff2",revision:"d5dd845698f1f91804004d3a74fb8aeb"},{url:"/_next/static/media/e58f9923a122d338-s.p.woff2",revision:"8ef41ccf9de5c6b1f1dd081670e93823"},{url:"/_next/static/media/facebook.ec3daa34.webp",revision:"c64ecb4743fc7a5c6aa4e9088e30dd4c"},{url:"/_next/static/media/facebook_circle.75465448.webp",revision:"d891ed7a868f6df38c6391dcb2c1ba24"},{url:"/_next/static/media/facebook_circle_mono.c4c0530f.webp",revision:"f93aa5bba995561d816683243ca93492"},{url:"/_next/static/media/facebook_mono.2123f1f3.webp",revision:"5dc2b7ba86cd92930f805e6fb74d39ac"},{url:"/_next/static/media/ff840cfebfb63b0c-s.p.woff2",revision:"302ec55f5b4320354ec6b35a53dead87"},{url:"/_next/static/media/instagram.f3acfbd2.webp",revision:"ce12832607b4138dd46d507d8832056b"},{url:"/_next/static/media/instagram_circle.a50b55a3.webp",revision:"798d201ca4096ab78e8bdbaa11f7778b"},{url:"/_next/static/media/instagram_circle_mono.84a09e86.webp",revision:"88fa7c52f1e2307ed514327e4edf772d"},{url:"/_next/static/media/instagram_mono.8106418d.webp",revision:"7739552abc9b647eb055640bbb6b3296"},{url:"/_next/static/media/linkedin.8e33860d.webp",revision:"c7cec304ec658405e1ba0684f82598da"},{url:"/_next/static/media/linkedin_circle.05ade0ca.webp",revision:"fc3a99ab78a78881715b901cc6f15268"},{url:"/_next/static/media/linkedin_circle_mono.631f184c.webp",revision:"0b9e9ca4c40cc5b7c2c3459c12d984a0"},{url:"/_next/static/media/linkedin_mono.23cdf21e.webp",revision:"9d114cd69175cb006f11b2b02c2c8788"},{url:"/_next/static/media/thread.2b1a37ad.webp",revision:"2ebc2f69e1b6c6d43ee0fef3642deeeb"},{url:"/_next/static/media/thread_circle.6e8b1f7e.webp",revision:"f935068d9afadda3d3b7026177aeb3d1"},{url:"/_next/static/media/thread_circle_mono.d15c8882.webp",revision:"baa9e226d867f1dbaf60897a04b1d209"},{url:"/_next/static/media/thread_mono.7785036c.webp",revision:"d1953b5651bb3ae28b87035133e62164"},{url:"/_next/static/media/tiktok.9243fbd9.webp",revision:"260257106d38409c5e716e2dceaf3fe0"},{url:"/_next/static/media/tiktok_circle.eb2c6fb2.webp",revision:"1ae38ee2f134290ffaca76d2e17c05af"},{url:"/_next/static/media/tiktok_mono.f8d07bac.webp",revision:"912f5a75e87fddae4040d4a33e25332e"},{url:"/_next/static/media/x.ae6955f1.webp",revision:"6d1325e40a5c426a243cf15d0092eb11"},{url:"/_next/static/media/x_circle.ab449725.webp",revision:"0f09f7f36d06efb4fc53dca431b4c000"},{url:"/_next/static/media/x_circle_mono.1af1a797.webp",revision:"8f6b0b9577597d6ea82d02f232cfab4e"},{url:"/_next/static/media/x_mono.8bfb9707.webp",revision:"e661f674533430884ae0efeaadb98bd1"},{url:"/_next/static/media/youtube.ec6c14e3.webp",revision:"b7b136dce0df84fe9b39f9f8a76faa4a"},{url:"/_next/static/media/youtube_circle.c98d5432.webp",revision:"09aa2981c00fb1bc3c3d99599c95f343"},{url:"/_next/static/media/youtube_mono.24874c3a.webp",revision:"4a17d2aabc029af75e927f18c036b6af"},{url:"/apple-touch-icon-precomposed.png",revision:"7f87b65f2c6502b8e6bebc7e43953f33"},{url:"/apple-touch-icon.png",revision:"b01685ae084bf0039f98b1a3484c466d"},{url:"/assets/images/background_default.webp",revision:"17f0be59db1bccbece3c3cb3c3d1ae3a"},{url:"/assets/images/background_green.webp",revision:"98f1610b7e03fea2a3b83e14b98b037c"},{url:"/assets/images/background_pink.webp",revision:"5710d584004e16f19e68832cd05827ea"},{url:"/assets/images/background_yellow.webp",revision:"9bf69ab06dae81275da71d38e3cfa9b0"},{url:"/assets/images/facebook-logo.svg",revision:"b2dd6d27d5787865946580275ac78781"},{url:"/assets/images/insta-logo.svg",revision:"e636cdc4ebe897630eb3102e1c5a168d"},{url:"/assets/images/logo-black.webp",revision:"07465ad2f1f0c189fb76965f294d9513"},{url:"/assets/images/logo.webp",revision:"072821d0e388ffccb5cdde989c853b63"},{url:"/assets/images/sns/facebook.webp",revision:"c64ecb4743fc7a5c6aa4e9088e30dd4c"},{url:"/assets/images/sns/facebook_circle.webp",revision:"d891ed7a868f6df38c6391dcb2c1ba24"},{url:"/assets/images/sns/facebook_circle_mono.webp",revision:"f93aa5bba995561d816683243ca93492"},{url:"/assets/images/sns/facebook_mono.webp",revision:"5dc2b7ba86cd92930f805e6fb74d39ac"},{url:"/assets/images/sns/index.ts",revision:"c02e02e2750b3a9b03d681d8c554ff60"},{url:"/assets/images/sns/instagram.webp",revision:"ce12832607b4138dd46d507d8832056b"},{url:"/assets/images/sns/instagram_circle.webp",revision:"798d201ca4096ab78e8bdbaa11f7778b"},{url:"/assets/images/sns/instagram_circle_mono.webp",revision:"88fa7c52f1e2307ed514327e4edf772d"},{url:"/assets/images/sns/instagram_mono.webp",revision:"7739552abc9b647eb055640bbb6b3296"},{url:"/assets/images/sns/linkedin.webp",revision:"c7cec304ec658405e1ba0684f82598da"},{url:"/assets/images/sns/linkedin_circle.webp",revision:"fc3a99ab78a78881715b901cc6f15268"},{url:"/assets/images/sns/linkedin_circle_mono.webp",revision:"0b9e9ca4c40cc5b7c2c3459c12d984a0"},{url:"/assets/images/sns/linkedin_mono.webp",revision:"9d114cd69175cb006f11b2b02c2c8788"},{url:"/assets/images/sns/thread.webp",revision:"2ebc2f69e1b6c6d43ee0fef3642deeeb"},{url:"/assets/images/sns/thread_circle.webp",revision:"f935068d9afadda3d3b7026177aeb3d1"},{url:"/assets/images/sns/thread_circle_mono.webp",revision:"baa9e226d867f1dbaf60897a04b1d209"},{url:"/assets/images/sns/thread_mono.webp",revision:"d1953b5651bb3ae28b87035133e62164"},{url:"/assets/images/sns/tiktok.webp",revision:"260257106d38409c5e716e2dceaf3fe0"},{url:"/assets/images/sns/tiktok_circle.webp",revision:"1ae38ee2f134290ffaca76d2e17c05af"},{url:"/assets/images/sns/tiktok_mono.webp",revision:"912f5a75e87fddae4040d4a33e25332e"},{url:"/assets/images/sns/x.webp",revision:"6d1325e40a5c426a243cf15d0092eb11"},{url:"/assets/images/sns/x_circle.webp",revision:"0f09f7f36d06efb4fc53dca431b4c000"},{url:"/assets/images/sns/x_circle_mono.webp",revision:"8f6b0b9577597d6ea82d02f232cfab4e"},{url:"/assets/images/sns/x_mono.webp",revision:"e661f674533430884ae0efeaadb98bd1"},{url:"/assets/images/sns/youtube.webp",revision:"b7b136dce0df84fe9b39f9f8a76faa4a"},{url:"/assets/images/sns/youtube_circle.webp",revision:"09aa2981c00fb1bc3c3d99599c95f343"},{url:"/assets/images/sns/youtube_mono.webp",revision:"4a17d2aabc029af75e927f18c036b6af"},{url:"/assets/images/threads-logo.svg",revision:"e67e8a408fd87a782ac68e9b03b3589c"},{url:"/assets/logo/logo-1024x1024.png",revision:"5c2304358ca643ab9c95b9dd6835875e"},{url:"/assets/logo/logo-192x192.png",revision:"d629a5edbc63ab13263041336c293ccd"},{url:"/assets/logo/logo-256x256.png",revision:"f08c11ba0b93aa5fb9d5d0be8f45ba5c"},{url:"/assets/logo/logo-512x512.png",revision:"e5c012cdb0a9df1807e7a4348fb3e072"},{url:"/assets/splash/4__iPhone_SE__iPod_touch_5th_generation_and_later_portrait.png",revision:"f6bea25ab83f5910f24f2f975bd17191"},{url:"/assets/splash/iPhone_11_Pro_Max__iPhone_XS_Max_portrait.png",revision:"0938d698ccd19b94a84daed03868aa7c"},{url:"/assets/splash/iPhone_11__iPhone_XR_portrait.png",revision:"ff6ba4860cef64674f90e8983f475003"},{url:"/assets/splash/iPhone_13_mini__iPhone_12_mini__iPhone_11_Pro__iPhone_XS__iPhone_X_portrait.png",revision:"d1f9294a341da1573c96598be7abea53"},{url:"/assets/splash/iPhone_14_Plus__iPhone_13_Pro_Max__iPhone_12_Pro_Max_portrait.png",revision:"a2f28fb6141e4cd653421676552fa2d3"},{url:"/assets/splash/iPhone_16_Plus__iPhone_15_Pro_Max__iPhone_15_Plus__iPhone_14_Pro_Max_portrait.png",revision:"896722e099fd2723e8b21a5bb9dc5754"},{url:"/assets/splash/iPhone_16_Pro_Max_portrait.png",revision:"f9039b39c6563f6a0548976d7561376b"},{url:"/assets/splash/iPhone_16_Pro_portrait.png",revision:"e3f8b73115d945ea55b4c4b842ea1733"},{url:"/assets/splash/iPhone_16__iPhone_15_Pro__iPhone_15__iPhone_14_Pro_portrait.png",revision:"d4125c02e406a497a8f100e5f0928476"},{url:"/assets/splash/iPhone_16e__iPhone_14__iPhone_13_Pro__iPhone_13__iPhone_12_Pro__iPhone_12_portrait.png",revision:"4366f0ed6955ae22ce4634f2ec88746b"},{url:"/assets/splash/iPhone_8_Plus__iPhone_7_Plus__iPhone_6s_Plus__iPhone_6_Plus_portrait.png",revision:"e2d176dd57efe8bb631536aea23eb655"},{url:"/assets/splash/iPhone_8__iPhone_7__iPhone_6s__iPhone_6__4.7__iPhone_SE_portrait.png",revision:"40a4ed04626b59e1b376976d93fd01b8"},{url:"/assets/svgs/google-logo.svg",revision:"f1e0e00aac37747a6c186405c21d77f7"},{url:"/assets/svgs/icon.svg",revision:"34913e04cf0a326bc37cf8f54fe046d5"},{url:"/fonts/CreatoDisplay-BoldItalic.woff2",revision:"d5dd845698f1f91804004d3a74fb8aeb"},{url:"/fonts/CreatoDisplay-ExtraBoldItalic.woff2",revision:"fae6df7521d3c7a4978234c4311ccb2b"},{url:"/fonts/CreatoDisplay-MediumItalic.woff2",revision:"100caa3008af382762a57d6e0e9d2673"},{url:"/fonts/CreatoDisplay-RegularItalic.woff2",revision:"8ef41ccf9de5c6b1f1dd081670e93823"},{url:"/fonts/PretendardVariable.woff2",revision:"302ec55f5b4320354ec6b35a53dead87"},{url:"/manifest.json",revision:"3ef2391913f8e6014f37e7b9b757aba0"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:i,state:a})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const s=e.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")}));
+if(!self.define){let e,s={};const a=(a,i)=>(a=new URL(a+".js",i).href,s[a]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=a,e.onload=s,document.head.appendChild(e)}else e=a,importScripts(a),s()})).then((()=>{let e=s[a];if(!e)throw new Error(`Module ${a} didn’t register its module`);return e})));self.define=(i,n)=>{const t=e||("document"in self?document.currentScript.src:"")||location.href;if(s[t])return;let c={};const r=e=>a(e,t),o={module:{uri:t},exports:c,require:r};s[t]=Promise.all(i.map((e=>o[e]||r(e)))).then((e=>(n(...e),c)))}}define(["./workbox-e9849328"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"3e87470c4ff80c14c5e8ce3281636708"},{url:"/_next/static/5X9bntb9590CKPmmAwgrh/_buildManifest.js",revision:"a9fb6c6c2b196342b346f13620ff8e60"},{url:"/_next/static/5X9bntb9590CKPmmAwgrh/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/151-fb9977fa774b1fb0.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/224-3864b7b3b6d20906.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/465-997053893adc2514.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/54-8b437171fa84de0a.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/705-b4a7d38f16abd233.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/723-a16144bda9e8eb17.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/890-ac3f3eed02577339.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/921-262cef18686f8597.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/997-82fa72b95b599b2a.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/app/(auth)/policy/page-5bd013a08f6a1476.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/app/(auth)/sign-in/page-5ba1e21b30685f75.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/app/(auth)/sign-up-complete/page-ef6b2d986eb8ba9b.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/app/(auth)/sign-up/page-fdfc2c9103518b54.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/app/(main)/main/page-3b95302d5c9e935d.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/app/(main)/new-content/layout-9e9b48d91c57b0a1.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/app/(main)/new-content/page-29a3c0c51f9b0f51.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/app/(modal)/warning/page-283ade49fe1204f5.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/app/@modal/(.)warning/page-bc741b249d8fc91f.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/app/@modal/default-faae72e4101ecadb.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/app/_not-found/page-c47ff0b24eac5816.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/app/error-cae605b8839498b7.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/app/layout-6f4bee16b1609f35.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/app/loading-bdcb4e5f0b2cb86d.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/app/not-found-8771b6be3a4e8aa2.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/app/onboarding/page-b4818e7829142eba.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/ca777b1b-1c1ec62045d69b42.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/framework-6e06c675866dc992.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/main-731a1afcd85d304c.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/main-app-063f3418fe6187ac.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/pages/_app-399de83ec01c3392.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/pages/_error-82c8287d732697b1.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-70b9a19db61b851d.js",revision:"5X9bntb9590CKPmmAwgrh"},{url:"/_next/static/css/291a1637afd87740.css",revision:"291a1637afd87740"},{url:"/_next/static/media/5d85a1edd96c32f7-s.p.woff2",revision:"fae6df7521d3c7a4978234c4311ccb2b"},{url:"/_next/static/media/9edda583dedd31eb-s.p.woff2",revision:"100caa3008af382762a57d6e0e9d2673"},{url:"/_next/static/media/e27282d13bd93057-s.p.woff2",revision:"d5dd845698f1f91804004d3a74fb8aeb"},{url:"/_next/static/media/e58f9923a122d338-s.p.woff2",revision:"8ef41ccf9de5c6b1f1dd081670e93823"},{url:"/_next/static/media/facebook.ec3daa34.webp",revision:"c64ecb4743fc7a5c6aa4e9088e30dd4c"},{url:"/_next/static/media/facebook_circle.75465448.webp",revision:"d891ed7a868f6df38c6391dcb2c1ba24"},{url:"/_next/static/media/facebook_circle_mono.c4c0530f.webp",revision:"f93aa5bba995561d816683243ca93492"},{url:"/_next/static/media/facebook_mono.2123f1f3.webp",revision:"5dc2b7ba86cd92930f805e6fb74d39ac"},{url:"/_next/static/media/ff840cfebfb63b0c-s.p.woff2",revision:"302ec55f5b4320354ec6b35a53dead87"},{url:"/_next/static/media/instagram.f3acfbd2.webp",revision:"ce12832607b4138dd46d507d8832056b"},{url:"/_next/static/media/instagram_circle.a50b55a3.webp",revision:"798d201ca4096ab78e8bdbaa11f7778b"},{url:"/_next/static/media/instagram_circle_mono.84a09e86.webp",revision:"88fa7c52f1e2307ed514327e4edf772d"},{url:"/_next/static/media/instagram_mono.8106418d.webp",revision:"7739552abc9b647eb055640bbb6b3296"},{url:"/_next/static/media/linkedin.8e33860d.webp",revision:"c7cec304ec658405e1ba0684f82598da"},{url:"/_next/static/media/linkedin_circle.05ade0ca.webp",revision:"fc3a99ab78a78881715b901cc6f15268"},{url:"/_next/static/media/linkedin_circle_mono.631f184c.webp",revision:"0b9e9ca4c40cc5b7c2c3459c12d984a0"},{url:"/_next/static/media/linkedin_mono.23cdf21e.webp",revision:"9d114cd69175cb006f11b2b02c2c8788"},{url:"/_next/static/media/thread.2b1a37ad.webp",revision:"2ebc2f69e1b6c6d43ee0fef3642deeeb"},{url:"/_next/static/media/thread_circle.6e8b1f7e.webp",revision:"f935068d9afadda3d3b7026177aeb3d1"},{url:"/_next/static/media/thread_circle_mono.d15c8882.webp",revision:"baa9e226d867f1dbaf60897a04b1d209"},{url:"/_next/static/media/thread_mono.7785036c.webp",revision:"d1953b5651bb3ae28b87035133e62164"},{url:"/_next/static/media/tiktok.9243fbd9.webp",revision:"260257106d38409c5e716e2dceaf3fe0"},{url:"/_next/static/media/tiktok_circle.eb2c6fb2.webp",revision:"1ae38ee2f134290ffaca76d2e17c05af"},{url:"/_next/static/media/tiktok_mono.f8d07bac.webp",revision:"912f5a75e87fddae4040d4a33e25332e"},{url:"/_next/static/media/x.ae6955f1.webp",revision:"6d1325e40a5c426a243cf15d0092eb11"},{url:"/_next/static/media/x_circle.ab449725.webp",revision:"0f09f7f36d06efb4fc53dca431b4c000"},{url:"/_next/static/media/x_circle_mono.1af1a797.webp",revision:"8f6b0b9577597d6ea82d02f232cfab4e"},{url:"/_next/static/media/x_mono.8bfb9707.webp",revision:"e661f674533430884ae0efeaadb98bd1"},{url:"/_next/static/media/youtube.ec6c14e3.webp",revision:"b7b136dce0df84fe9b39f9f8a76faa4a"},{url:"/_next/static/media/youtube_circle.c98d5432.webp",revision:"09aa2981c00fb1bc3c3d99599c95f343"},{url:"/_next/static/media/youtube_mono.24874c3a.webp",revision:"4a17d2aabc029af75e927f18c036b6af"},{url:"/apple-touch-icon-precomposed.png",revision:"7f87b65f2c6502b8e6bebc7e43953f33"},{url:"/apple-touch-icon.png",revision:"b01685ae084bf0039f98b1a3484c466d"},{url:"/assets/images/background_default.webp",revision:"17f0be59db1bccbece3c3cb3c3d1ae3a"},{url:"/assets/images/background_green.webp",revision:"98f1610b7e03fea2a3b83e14b98b037c"},{url:"/assets/images/background_pink.webp",revision:"5710d584004e16f19e68832cd05827ea"},{url:"/assets/images/background_yellow.webp",revision:"9bf69ab06dae81275da71d38e3cfa9b0"},{url:"/assets/images/facebook-logo.svg",revision:"b2dd6d27d5787865946580275ac78781"},{url:"/assets/images/insta-logo.svg",revision:"e636cdc4ebe897630eb3102e1c5a168d"},{url:"/assets/images/logo-black.webp",revision:"07465ad2f1f0c189fb76965f294d9513"},{url:"/assets/images/logo.webp",revision:"072821d0e388ffccb5cdde989c853b63"},{url:"/assets/images/sns/facebook.webp",revision:"c64ecb4743fc7a5c6aa4e9088e30dd4c"},{url:"/assets/images/sns/facebook_circle.webp",revision:"d891ed7a868f6df38c6391dcb2c1ba24"},{url:"/assets/images/sns/facebook_circle_mono.webp",revision:"f93aa5bba995561d816683243ca93492"},{url:"/assets/images/sns/facebook_mono.webp",revision:"5dc2b7ba86cd92930f805e6fb74d39ac"},{url:"/assets/images/sns/index.ts",revision:"c02e02e2750b3a9b03d681d8c554ff60"},{url:"/assets/images/sns/instagram.webp",revision:"ce12832607b4138dd46d507d8832056b"},{url:"/assets/images/sns/instagram_circle.webp",revision:"798d201ca4096ab78e8bdbaa11f7778b"},{url:"/assets/images/sns/instagram_circle_mono.webp",revision:"88fa7c52f1e2307ed514327e4edf772d"},{url:"/assets/images/sns/instagram_mono.webp",revision:"7739552abc9b647eb055640bbb6b3296"},{url:"/assets/images/sns/linkedin.webp",revision:"c7cec304ec658405e1ba0684f82598da"},{url:"/assets/images/sns/linkedin_circle.webp",revision:"fc3a99ab78a78881715b901cc6f15268"},{url:"/assets/images/sns/linkedin_circle_mono.webp",revision:"0b9e9ca4c40cc5b7c2c3459c12d984a0"},{url:"/assets/images/sns/linkedin_mono.webp",revision:"9d114cd69175cb006f11b2b02c2c8788"},{url:"/assets/images/sns/thread.webp",revision:"2ebc2f69e1b6c6d43ee0fef3642deeeb"},{url:"/assets/images/sns/thread_circle.webp",revision:"f935068d9afadda3d3b7026177aeb3d1"},{url:"/assets/images/sns/thread_circle_mono.webp",revision:"baa9e226d867f1dbaf60897a04b1d209"},{url:"/assets/images/sns/thread_mono.webp",revision:"d1953b5651bb3ae28b87035133e62164"},{url:"/assets/images/sns/tiktok.webp",revision:"260257106d38409c5e716e2dceaf3fe0"},{url:"/assets/images/sns/tiktok_circle.webp",revision:"1ae38ee2f134290ffaca76d2e17c05af"},{url:"/assets/images/sns/tiktok_mono.webp",revision:"912f5a75e87fddae4040d4a33e25332e"},{url:"/assets/images/sns/x.webp",revision:"6d1325e40a5c426a243cf15d0092eb11"},{url:"/assets/images/sns/x_circle.webp",revision:"0f09f7f36d06efb4fc53dca431b4c000"},{url:"/assets/images/sns/x_circle_mono.webp",revision:"8f6b0b9577597d6ea82d02f232cfab4e"},{url:"/assets/images/sns/x_mono.webp",revision:"e661f674533430884ae0efeaadb98bd1"},{url:"/assets/images/sns/youtube.webp",revision:"b7b136dce0df84fe9b39f9f8a76faa4a"},{url:"/assets/images/sns/youtube_circle.webp",revision:"09aa2981c00fb1bc3c3d99599c95f343"},{url:"/assets/images/sns/youtube_mono.webp",revision:"4a17d2aabc029af75e927f18c036b6af"},{url:"/assets/images/threads-logo.svg",revision:"e67e8a408fd87a782ac68e9b03b3589c"},{url:"/assets/logo/logo-1024x1024.png",revision:"5c2304358ca643ab9c95b9dd6835875e"},{url:"/assets/logo/logo-192x192.png",revision:"d629a5edbc63ab13263041336c293ccd"},{url:"/assets/logo/logo-256x256.png",revision:"f08c11ba0b93aa5fb9d5d0be8f45ba5c"},{url:"/assets/logo/logo-512x512.png",revision:"e5c012cdb0a9df1807e7a4348fb3e072"},{url:"/assets/splash/4__iPhone_SE__iPod_touch_5th_generation_and_later_portrait.png",revision:"f6bea25ab83f5910f24f2f975bd17191"},{url:"/assets/splash/iPhone_11_Pro_Max__iPhone_XS_Max_portrait.png",revision:"0938d698ccd19b94a84daed03868aa7c"},{url:"/assets/splash/iPhone_11__iPhone_XR_portrait.png",revision:"ff6ba4860cef64674f90e8983f475003"},{url:"/assets/splash/iPhone_13_mini__iPhone_12_mini__iPhone_11_Pro__iPhone_XS__iPhone_X_portrait.png",revision:"d1f9294a341da1573c96598be7abea53"},{url:"/assets/splash/iPhone_14_Plus__iPhone_13_Pro_Max__iPhone_12_Pro_Max_portrait.png",revision:"a2f28fb6141e4cd653421676552fa2d3"},{url:"/assets/splash/iPhone_16_Plus__iPhone_15_Pro_Max__iPhone_15_Plus__iPhone_14_Pro_Max_portrait.png",revision:"896722e099fd2723e8b21a5bb9dc5754"},{url:"/assets/splash/iPhone_16_Pro_Max_portrait.png",revision:"f9039b39c6563f6a0548976d7561376b"},{url:"/assets/splash/iPhone_16_Pro_portrait.png",revision:"e3f8b73115d945ea55b4c4b842ea1733"},{url:"/assets/splash/iPhone_16__iPhone_15_Pro__iPhone_15__iPhone_14_Pro_portrait.png",revision:"d4125c02e406a497a8f100e5f0928476"},{url:"/assets/splash/iPhone_16e__iPhone_14__iPhone_13_Pro__iPhone_13__iPhone_12_Pro__iPhone_12_portrait.png",revision:"4366f0ed6955ae22ce4634f2ec88746b"},{url:"/assets/splash/iPhone_8_Plus__iPhone_7_Plus__iPhone_6s_Plus__iPhone_6_Plus_portrait.png",revision:"e2d176dd57efe8bb631536aea23eb655"},{url:"/assets/splash/iPhone_8__iPhone_7__iPhone_6s__iPhone_6__4.7__iPhone_SE_portrait.png",revision:"40a4ed04626b59e1b376976d93fd01b8"},{url:"/assets/svgs/google-logo.svg",revision:"f1e0e00aac37747a6c186405c21d77f7"},{url:"/assets/svgs/icon.svg",revision:"34913e04cf0a326bc37cf8f54fe046d5"},{url:"/fonts/CreatoDisplay-BoldItalic.woff2",revision:"d5dd845698f1f91804004d3a74fb8aeb"},{url:"/fonts/CreatoDisplay-ExtraBoldItalic.woff2",revision:"fae6df7521d3c7a4978234c4311ccb2b"},{url:"/fonts/CreatoDisplay-MediumItalic.woff2",revision:"100caa3008af382762a57d6e0e9d2673"},{url:"/fonts/CreatoDisplay-RegularItalic.woff2",revision:"8ef41ccf9de5c6b1f1dd081670e93823"},{url:"/fonts/PretendardVariable.woff2",revision:"302ec55f5b4320354ec6b35a53dead87"},{url:"/manifest.json",revision:"3ef2391913f8e6014f37e7b9b757aba0"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:a,state:i})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const s=e.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")}));
