+++ src/app/(auth)/sign-in/_components/SignInForm.tsx
@@ -5,8 +5,7 @@ import { z } from "zod";
 import { zodResolver } from "@hookform/resolvers/zod";
 import Link from "next/link";
 import { useToast } from "@/utils/useToast";
-import { Button } from "@/components/common/button";
-import { Input } from "@/components/common/input";
+import { Button, Input } from "@/components";
 import { emailCheckSchema } from "@/lib/zod";
 import { authService } from "@/lib/api/service";
 import { useRouter } from "next/navigation";
