+++ src/app/(main)/new-content/_components/common/index.ts
@@ -1,0 +1,2 @@
+export { default as InfoMainPlatform } from "./InfoMainPlatform";
+export { default as ProgressBar } from "./ProgressBar";
