--- src/app/(main)/new-content/_components/content-make/index.ts
@@ -1,1 +1,0 @@
-export { default as Bedge } from "./Bedge";
