--- src/app/(main)/new-content/_components/platform-select/index.ts
@@ -1,5 +1,0 @@
-export { default as HeaderContent } from "./HeaderContent";
-export { default as PlatformList } from "./PlatformList";
-export { default as Platform } from "./Platform";
-export { default as CategoryButton } from "./CategoryButton";
-export { default as Continue } from "./Continue";
