+++ src/app/(main)/new-content/_components/step/content-make/index.ts
@@ -1,0 +1,2 @@
+export { default as Bedge } from "./Bedge";
+export { default as ContentMakeStep } from "./ContentMakeStep";
