+++ src/app/(main)/new-content/_components/step/index.ts
@@ -1,0 +1,4 @@
+export * from "./platform-select";
+export * from "./content-make";
+export * from "./upload-confirm";
+export * from "./reserve";
