+++ src/app/(main)/new-content/_components/step/platform-select/Continue.tsx
