+++ src/app/(main)/new-content/_components/step/platform-select/index.ts
@@ -1,0 +1,7 @@
+export { default as Continue } from "./Continue";
+export { default as HeaderContent } from "./HeaderContent";
+export { default as PlatformList } from "./PlatformList";
+export { default as BottomContent } from "./BottomContent";
+export { default as PlatformSelectStep } from "./PlatformSelectStep";
+export { default as CategoryButton } from "./CategoryButton";
+export { default as Platform } from "./Platform";
