+++ src/app/(main)/new-content/_components/step/reserve/index.ts
@@ -1,0 +1,1 @@
+export { default as ReserveStep } from "./ReserveStep";
