+++ src/app/(main)/new-content/_components/step/upload-confirm/UploadConfirmStep.tsx
