+++ src/app/(main)/new-content/_components/step/upload-confirm/index.ts
@@ -1,0 +1,1 @@
+export { default as UploadConfirmStep } from "./UploadConfirmStep";
