--- src/app/(main)/new-content/_components/steps/index.ts
@@ -1,4 +1,0 @@
-export { default as PlatformSelectStep } from "./PlatformSelectStep";
-export { default as ContentMakeStep } from "./ContentMakeStep";
-export { default as ReserveStep } from "./ReserveStep";
-export { default as UploadConfirmStep } from "./UploadConfirmStep";
