+++ src/app/(main)/new-content/_components/store/index.ts
@@ -1,0 +1,2 @@
+export * from "./platform-store";
+export * from "./sheet-store";
