+++ src/app/(main)/new-content/layout.tsx
@@ -1,4 +1,4 @@
-import ProgressBar from "./_components/ProgressBar";
+import ProgressBar from "./_components/common/ProgressBar";
 
 export default async function NewContentLayout({ children }: { children: React.ReactNode }) {
   return (
