+++ src/app/onboarding/_components/OnBoardingFooter.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Button } from "@/components/common/button";
+import { Button } from "@/components";
 import { useRouter } from "next/navigation";
 
 const OnBoardingFooter = () => {
