+++ src/components/common/button/index.tsx
@@ -3,7 +3,7 @@ import { Slot } from "@radix-ui/react-slot";
 import { cva, type VariantProps } from "class-variance-authority";
 import cn from "@/utils/cn";
 
-const buttonVariants = cva("h-[50px] flex items-center justify-center gap-2 text-button1", {
+export const buttonVariants = cva("h-[50px] flex items-center justify-center gap-2 text-button1", {
   variants: {
     variant: {
       active: "bg-blue-blueblack text-grayscale-100",
@@ -50,4 +50,4 @@ const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
 );
 Button.displayName = "Button";
 
-export { Button, buttonVariants };
+export default Button;
