+++ src/components/common/input/index.tsx
@@ -78,4 +78,4 @@ const Input = React.forwardRef<
 );
 Input.displayName = "Input";
 
-export { Input };
+export default Input;
