+++ src/components/index.ts
@@ -1,1 +1,3 @@
 export * from "./common";
+export * from "./animation";
+export * from "./icons";
